+++ examples/tracing/jaeger/go/rideshare/rideshare/rideshare.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/pyroscope-io/client/pyroscope"
-	"github.com/pyroscope-io/otelpyroscope"
+	otelpyroscope "github.com/pyroscope-io/otel-profiling-go"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/jaeger"
 	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"