+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/RightClickIcon.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+
+export default function RightClickIcon() {
+  return (
+    <svg
+      id="Layer_2"
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 16.84 25.56"
+    >
+      <defs />
+      <g id="_">
+        <g>
+          <path
+            style={{ fill: '#7d7d7f' }}
+            d="M.38,16.25c0,5.14,3.07,9.31,6.86,9.31h2.74c3.79,0,6.86-4.17,6.86-9.31v-7.45H.38v7.45Z"
+          />
+          <path
+            style={{
+              fill: '#7d7d7f',
+              stroke: '#7d7d7f',
+              strokeMiterlimit: 10,
+              strokeWidth: '.75px',
+            }}
+            d="M7.92,.38h-.69C3.45,.38,.38,3.09,.38,6.43v1.21H7.92V.38Z"
+          />
+          <path
+            style={{ fill: '#020202' }}
+            d="M9.98,.38h-.69V7.65h7.54v-1.21C16.84,3.09,13.77,.38,9.98,.38Z"
+          />
+        </g>
+      </g>
+    </svg>
+  );
+}