+++ webapp/javascript/components/ExportData.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // import { SimpleSingle as TestData } from '@utils/testData';
 import { render, screen } from '@testing-library/react';
-import { Profile } from '@pyroscope/models';
+import { Profile } from '@pyroscope/models/src';
 import ExportData, { getFilename } from './ExportData';
 
 describe('ExportData', () => {